refactor(testability): rename function to get testability
Previously, getting testability was `window.angular2.getTestability`
This was because the plan was to export the API to the window as
angular2. However, the decision was changed to make this just `angular`
in https://github.com/angular/angular/commit/3177576ad6889fd7cb592d27ff0cd5c0d85bd9d6

To decouple testability from the rest of the Angular API, just make it
one function, `window.getAngularTestability`.