ci: re-enable chrome mobile tests on android (#32447)
With 63dff9c888c2c89afa7c8aff9747022768b42370, we temporarily disabled
the chrome mobile tests on Android because the launched browsers weren't
able to load the karma page through the tunnel.

The Saucelabs support responded to a ticket that we opened and said that
there were issues with those browsers which are now resolved.

PR Close #32447